internal_link [nfc]: Move narrowLink out of compose into a new internal_link file
